Add DailyUserProfile component and integrate it into SocialDailyPage and SocialFollowingPage
