+++ frontend/src/pages/SocialDailyPage.tsx
@@ -1,30 +1,18 @@
 import { useEffect, useState } from "react"
 import { useLocation, useParams } from "react-router-dom"
-import { Link } from "react-router-dom"
 import { useNavigate } from "react-router-dom"
 
-import { useQuery } from "@tanstack/react-query"
-import styled from "styled-components"
-
-import { LoaderCircleFull } from "@components/common/LoaderCircle"
-import PageBack from "@components/common/PageBack"
-import PageTitle from "@components/common/PageTitle"
+import DailyUserProfile from "@components/social/DailyUserProfile"
 import DateBar from "@components/social/common/DateBar"
-import { SimpleProfileImg } from "@components/social/common/SimpleProfile"
 import LogDetails from "@components/social/logDetails/LogDetails"
 
-import { getUserByUsername } from "@api/users.api"
-
 import { DateTime } from "luxon"
-import { useTranslation } from "react-i18next"
-import { toast } from "react-toastify"
 
 const SocialDailyPage = () => {
     const navigate = useNavigate()
     const { username: usernameWithAt } = useParams()
     const location = useLocation()
     const { selectedDate: receivedDate } = location.state || {}
-    const { t } = useTranslation("translation")
 
     useEffect(() => {
         if (usernameWithAt!.at(0) !== "@") {
@@ -34,21 +22,6 @@ const SocialDailyPage = () => {
 
     const username = usernameWithAt!.slice(1)
 
-    const {
-        data: user,
-        isPending,
-        isError,
-    } = useQuery({
-        queryKey: ["users", username],
-        queryFn: () => getUserByUsername(username),
-    })
-
-    useEffect(() => {
-        if (isError) {
-            toast.error(t("common.error_load"))
-        }
-    }, [isError])
-
     const [selectedDate, setSelectedDate] = useState(
         receivedDate ||
             DateTime.now()
@@ -56,77 +29,20 @@ const SocialDailyPage = () => {
                 .toISO(),
     )
 
-    if (isPending || isError) {
-        return (
-            <>
-                <Header>
-                    <HeaderTexts>
-                        <PageBack defaultTo="/app/social" />
-                    </HeaderTexts>
-                </Header>
-
-                <LoaderCircleFull />
-            </>
-        )
-    }
-
-    const userPagePath = `/app/users/@${user.username}`
-
     return (
         <>
-            <Header>
-                <HeaderTexts>
-                    <PageBack defaultTo="/app/social" />
-                    <Link to={userPagePath}>
-                        <UsernameTitle $cursor="pointer">
-                            {user.username}
-                        </UsernameTitle>
-                    </Link>
-                </HeaderTexts>
-
-                <ProfileWrapper to={userPagePath}>
-                    <SimpleProfileImg src={user.profile_img} />
-                </ProfileWrapper>
-            </Header>
-
+            <DailyUserProfile username={username} back />
             <DateBar
                 selectedDate={selectedDate}
                 setSelectedDate={setSelectedDate}
             />
-
             <LogDetails
-                username={user.username}
+                username={username}
                 selectedDate={selectedDate}
                 displayProfile={false}
             />
         </>
     )
 }
-
-const Header = styled.div`
-    width: 100%;
-
-    display: flex;
-    justify-content: space-between;
-`
-
-const HeaderTexts = styled.div`
-    width: 70%;
-
-    display: flex;
-    flex-direction: column;
-    justify-content: left;
-`
-
-const UsernameTitle = styled(PageTitle)`
-    overflow-x: clip;
-    text-overflow: ellipsis;
-
-    white-space: nowrap;
-`
-
-const ProfileWrapper = styled(Link)`
-    width: 4rem;
-`
 
 export default SocialDailyPage
